// src/app/models/projects.ts
export interface Project {
  id: string;
  img: string;
  title: string;
  personName: string;
  discription: string;
  budget: string;
  durationproject: string;
  numberoffers?: string;
  offers?: string;
  publishData: string;
  time: string;
  Averageoffers?: string;
  skills?: string;
  open: boolean;
  isFeatured: boolean;
  isApproved: boolean;
  accepted: boolean;
  country: string;
  catId: string;
  catergory?: string;
}